+++ data/stories.json
@@ -1,0 +1,11 @@
+[{
+    "images": [
+        "images\\image6.jpg",
+        "images\\image7.jpg",
+        "images\\image8.jpg"
+    ],
+    "contentId": "S1",
+    "authorId": "authorId",
+    "content": "wrw",
+    "timestamp": "2024-12-05 20:27:06"
+}]
