+++ src/main/java/backend/contentCreation/IContent.java
@@ -1,4 +1,4 @@
-package backend;
+package backend.contentCreation;
 import org.json.JSONArray;
 
 import java.util.List;
